Refactor package structure for yaml properties provider
